+++ Changelog.md
@@ -1,6 +1,7 @@
 Unreleased
 -------------------------
 
+* [Enhancement] Add Mac Cmd+V paste support for the lab terminal.
 * [Bug fix] Rename the CSS class for labs from `hastexblock` to `stackamolexblock`, post the XBlock rename.
 
 Version 9.0.0 (2026-05-13)